+++ coinbase/http.request.js
@@ -2,7 +2,7 @@ const http = require('https');
 const log = require('../lib/log');
 module.exports = params => {
   const makeRequest = function (resolve, reject) {
-    log.debug({ params });
+    // log.debug({ params });
     params.resolve = resolve;
     params.reject = reject;
     const req = http.request(params, function (res) {