+++ test/nock/get.order.404.js
@@ -3,5 +3,5 @@ const nock = require('nock');
 
 module.exports = nock(config.api)
   .persist()
-  .get('/api/v3/brokerage/orders/404')
+  .get('/api/v3/brokerage/orders/historical/404')
   .reply(404, {});