+++ test/nock/get.order.netfail.js
@@ -3,5 +3,5 @@ const nock = require('nock');
 
 module.exports = nock(config.api)
   .persist()
-  .get('/api/v3/brokerage/orders/fail')
+  .get('/api/v3/brokerage/orders/historical/fail')
   .replyWithError('ENOTFOUND');