+++ tools/create.history.js
@@ -52,14 +52,14 @@ log.ok(`backed up history file in ${backup}`);
     //   log.debug('no volume?', f);
     // }
     const funds = multiply(f.price, f.size);
-    const Funds = f.side === 'sell' ? multiply(funds, -1) : funds;
+    const Funds = f.side === 'SELL' ? multiply(funds, -1) : funds;
     return {
       Time: f.created_at,
       Price: f.price,
       Holding: 0,
       Value: 0,
       Funds,
-      Shares: f.side === 'sell' ? multiply(f.size, -1) : f.size,
+      Shares: f.side === 'SELL' ? multiply(f.size, -1) : f.size,
       PeriodRate: 0,
       ExpectedGain: 0,
       TotalInput: Funds,
@@ -71,7 +71,7 @@ log.ok(`backed up history file in ${backup}`);
       Liquid: 0,
       Profit: 0,
       ID: f.order_id,
-      Type: 'market',
+      Type: 'MARKET',
       Method: 'cron',
       TargetAPY: config.apy,
     };